requirements.txt
google-generativeai==0.8.5
streamlit==1.45.1
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
pandas
plotly
